Identify Debian system only with the dist part.
The code in the osdetect module remains small, but there is no garranty that
the platform module don't change in the future and the patch doesn't by
applied. In favor this only need one single part to be declared in the
src/python24.cfg and could works in new Debian releases.